requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.0.1
Flask-Migrate==4.0.7
flask-restx==1.1.0